clean-up drawing processors
