+++ src/ImageSharp.Drawing/ImageSharp.Drawing.csproj
@@ -40,9 +40,9 @@
 
   <ItemGroup>
     <AdditionalFiles Include="..\..\standards\stylecop.json" />
+    <PackageReference Include="StyleCop.Analyzers" Version="1.1.118" PrivateAssets="All" />
     <PackageReference Include="SixLabors.Fonts" Version="1.0.0-beta0008" />
     <PackageReference Include="SixLabors.Shapes" Version="1.0.0-beta0008" />
-    <PackageReference Include="StyleCop.Analyzers" Version="1.1.1-rc.114" PrivateAssets="All" />
   </ItemGroup>
 
   <PropertyGroup>
