+++ src/ImageSharp.Drawing/Processing/Processors/Drawing/FillProcessor.cs
@@ -1,28 +1,38 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using System;
-using System.Buffers;
 using System.Threading.Tasks;
-using SixLabors.ImageSharp.Advanced;
-using SixLabors.ImageSharp.Memory;
-using SixLabors.ImageSharp.ParallelUtils;
+
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.Memory;
-using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Processing.Processors.Drawing
 {
+    /// <summary>
+    /// Defines a processor to fill an <see cref="Image"/> with the given <see cref="IBrush"/>
+    /// using blending defined by the given <see cref="GraphicsOptions"/>.
+    /// </summary>
     public class FillProcessor : IImageProcessor
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="FillProcessor"/> class.
+        /// </summary>
+        /// <param name="brush">The brush to use for filling.</param>
+        /// <param name="options">The <see cref="GraphicsOptions"/> defining how to blend the brush pixels over the image pixels.</param>
         public FillProcessor(IBrush brush, GraphicsOptions options)
         {
             this.Brush = brush;
             this.Options = options;
         }
 
+        /// <summary>
+        /// Gets the <see cref="IBrush"/> used for filling the destination image.
+        /// </summary>
         public IBrush Brush { get; }
-        
+
+        /// <summary>
+        /// Gets the <see cref="GraphicsOptions"/> defining how to blend the brush pixels over the image pixels.
+        /// </summary>
         public GraphicsOptions Options { get; }
 
         /// <inheritdoc />
@@ -30,110 +40,6 @@ namespace SixLabors.ImageSharp.Processing.Processors.Drawing
             where TPixel : struct, IPixel<TPixel>
         {
             return new FillProcessor<TPixel>(this);
-        }
-    }
-    
-    /// <summary>
-    /// Using the brush as a source of pixels colors blends the brush color with source.
-    /// </summary>
-    /// <typeparam name="TPixel">The pixel format.</typeparam>
-    internal class FillProcessor<TPixel> : ImageProcessor<TPixel>
-        where TPixel : struct, IPixel<TPixel>
-    {
-        private readonly FillProcessor definition;
-
-        public FillProcessor(FillProcessor definition)
-        {
-            this.definition = definition;
-        }
-
-        /// <inheritdoc/>
-        protected override void OnFrameApply(ImageFrame<TPixel> source, Rectangle sourceRectangle, Configuration configuration)
-        {
-            int startX = sourceRectangle.X;
-            int endX = sourceRectangle.Right;
-            int startY = sourceRectangle.Y;
-            int endY = sourceRectangle.Bottom;
-
-            // Align start/end positions.
-            int minX = Math.Max(0, startX);
-            int maxX = Math.Min(source.Width, endX);
-            int minY = Math.Max(0, startY);
-            int maxY = Math.Min(source.Height, endY);
-
-            int width = maxX - minX;
-
-            var workingRect = Rectangle.FromLTRB(minX, minY, maxX, maxY);
-
-            IBrush brush = this.definition.Brush;
-            GraphicsOptions options = this.definition.Options;
-
-            // If there's no reason for blending, then avoid it.
-            if (this.IsSolidBrushWithoutBlending(out SolidBrush solidBrush))
-            {
-                ParallelExecutionSettings parallelSettings = configuration.GetParallelSettings().MultiplyMinimumPixelsPerTask(4);
-
-                TPixel colorPixel = solidBrush.Color.ToPixel<TPixel>();
-                
-                ParallelHelper.IterateRows(
-                    workingRect,
-                    parallelSettings,
-                    rows =>
-                        {
-                            for (int y = rows.Min; y < rows.Max; y++)
-                            {
-                                source.GetPixelRowSpan(y).Slice(minX, width).Fill(colorPixel);
-                            }
-                        });
-            }
-            else
-            {
-                // Reset offset if necessary.
-                if (minX > 0)
-                {
-                    startX = 0;
-                }
-
-                if (minY > 0)
-                {
-                    startY = 0;
-                }
-
-                using (IMemoryOwner<float> amount = source.MemoryAllocator.Allocate<float>(width))
-                using (BrushApplicator<TPixel> applicator = brush.CreateApplicator(
-                    source,
-                    sourceRectangle,
-                    options))
-                {
-                    amount.GetSpan().Fill(1f);
-
-                    ParallelHelper.IterateRows(
-                        workingRect,
-                        configuration,
-                        rows =>
-                            {
-                                for (int y = rows.Min; y < rows.Max; y++)
-                                {
-                                    int offsetY = y - startY;
-                                    int offsetX = minX - startX;
-
-                                    applicator.Apply(amount.GetSpan(), offsetX, offsetY);
-                                }
-                            });
-                }
-            }
-        }
-
-        private bool IsSolidBrushWithoutBlending(out SolidBrush solidBrush)
-        {
-            solidBrush = this.definition.Brush as SolidBrush;
-
-            if (solidBrush == null)
-            {
-                return false;
-            }
-
-            return this.definition.Options.IsOpaqueColorWithoutBlending(solidBrush.Color);
         }
     }
 }
